requirements.txt
requests
python-dateutil
azure-keyvault-secrets
azure-identity
azure-storage-blob
azure-monitor-query
pandas
